refactor: change the return type of str_has() & str_pat() from int to bool
The functions in Libdao which return int
returns 0 for successes and non-0 for failures.

However, these functions return 1 when the match succeeded (matched)
and 0 when the match failed (not matched),
violating the convention of Libdao.

By change their return type to bool, the maintainers can more easily spot
that these functions return true on success and false otherwise,
which seems more intuitive.